combine r50 nvme to nvme.py
